reduced data width of encoder vars

+++ src/drv/encoder_driver.cpp
@@ -1,8 +1,8 @@
 #include "encoder_driver.h"
 #include <Arduino.h>
 
-volatile static uint64_t er_count = 0;
-volatile static uint64_t el_count = 0;
+volatile static uint32_t er_count = 0;
+volatile static uint32_t el_count = 0;
 
 static void er1_int_handler(void){
     er_count++;
@@ -29,10 +29,10 @@ void encoder_setup_drv(uint8_t er1, uint8_t er2, uint8_t el1, uint8_t el2){
     attachInterrupt(digitalPinToInterrupt(el2), el2_int_handler, RISING);
 };
 
-uint64_t counts_left_drv(void) {
+uint32_t counts_left_drv(void) {
     return el_count;
 }
 
-uint64_t counts_right_drv(void){
+uint32_t counts_right_drv(void){
     return er_count;
 }
